+++ plugins/grideye_mem_read.c
@@ -39,6 +39,7 @@ static const struct grideye_plugin_api api = {
     NULL,            /* input format */
     "xml",            /* output format */
     NULL,
+    NULL,
     mem_read_test,
     NULL
 };