+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Grideye_agent Changelog
 
+* Added getopt to plugin API
+* Added synamic yangmetrics to callhome
+* Added error check for http_callhome and http_data
 * Hardening after AFL and valgrind runs
 * Added curl-timeout (-T) with default value 1min
 * For http controller communication, added resiliency for curl couldn't connect, timeout and bad gateway.